refactor(labnative): breaking logic in some files
